rename ardour strips (only default values yet)
Currently, the local channels are labelled L01, L02, etc and the channels received by tpf-client
are labeled 01, 02, 03, etc. The idea is that channels can be renamed and the names are reflected in
the ardour instances of the peers. The names keep the prefix, but allow to append any string. Example:
  02_BassFlute